EL SUPERCOMMIT, leer descripción
1. SFX_NINJA_REVEAL
2. Bajarle la velocidad a la Pata (peor enemigo del juego)
3. Arreglar bug en el que el láser del tiralapiceras (no el tiralapiceras en sí) buscaba al Ninja cuando estaba invisible
4. Algunos cambios en el sonido
5. MessageBox weplayista
6. Cambios menores/fixes en EnemySpawner, CustomRangeFloat y PauseScript*
6. Hacer que se puedan usar los cheats en el compilado apretando F3**
∞-1. Aprendí a usar `get` y `set` bien después de dos años de utilizar este lenguaje
∞. Seguro hice algo más que me olvidé de poner acá

* bueno no tan menor porque cambié todo el script para que sea un singleton pero bue

** siendo la fecha que es, no hice un GameManager. Lo metí directo en el singleton de la MessageBox. Que D's me perdone.

+++ Assets/Other Scripts/CustomRangeFloat.cs
@@ -7,8 +7,8 @@ using UnityEngine;
 [System.Serializable]
 public class CustomRangeFloat : ISerializationCallbackReceiver
 {
-    private int min;
-    private int max;
+    private float min;
+    private float max;
     [SerializeField] private float value;
 
     public float Value
@@ -17,7 +17,7 @@ public class CustomRangeFloat : ISerializationCallbackReceiver
         set => this.value = Mathf.Clamp(value, min, max);
     }
 
-    public CustomRangeFloat(int minIncluive, int maxIncluive, float initialValue)
+    public CustomRangeFloat(float minIncluive, float maxIncluive, float initialValue)
     {
         this.min = minIncluive;
         this.max = maxIncluive;
